frontend: load tasks after successful login
